解决了一些warning TODO ：libmpg123。lib 有众多warning需要解决。。

+++ demotry/CFlacDecoder.cpp
@@ -188,7 +188,7 @@ ARESULT CFlacDecoder::Close()
 
 int CFlacDecoder::GetFullTime()
 {
-	return m_dSampleCount / m_StreamInfo.data.stream_info.sample_rate  * 1000;
+	return int(m_dSampleCount / m_StreamInfo.data.stream_info.sample_rate  * 1000);
 }
 
 int CFlacDecoder::GetCurTime()
